Pseudocode for board panel
Set up pseudocode for the board panel, also made it so that when first rendered, the board panel will draw the board. This is temporary, and will only remain in effect until the program works well enough so that the board is re-drawn appropriately. (Also replaced the ui.drawboard calls with gui.drawboard)